@@ -32,9 +32,9 @@
   "devDependencies": {
     "@formily/reactive": "^2.2.16",
     "@formily/reactive-react": "^2.2.16",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "react": "^17.0.2"
+    "@types/react": "^18.2.48",
+    "@types/react-dom": "^18.2.18",
+    "react": "^18.2.0"
   },
   "peerDependencies": {
     "@formily/reactive": "^2.2.16",
